requirements.txt
psycopg2
django-bootstrap3
django-widget-tweaks
django-bootstrap-dynamic-formsets
celery
rpy2
django-tables2
django-filter
xlwt
scipy
colour
numpy
sklearn
